+++ Tharunpp/__init__.py
@@ -1,0 +1,13 @@
+"""Tharun++"""
+
+from . import core, cli
+
+__all__ = ["core", "cli"]
+
+__version__ = "1.0.0"
+__version_str__ = (
+    f"Tharun++ v{__version__}. "
+    + "A programming language powered by Tamil cinema comedy dialogues. "
+    + "Created by Tharun Kumar."
+)
+