+++ Tharunpp/cli/__init__.py
@@ -1,0 +1,5 @@
+"""Tharun++ CLI Module"""
+
+from .main import app
+
+__all__ = ['app']